improved ex01

+++ cpp_module_08/ex01/include/Span.hpp
@@ -43,8 +43,8 @@ public:
 
 
 private:
-    std::vector<int>    _numbers;
-    uint        _max_size;
+    std::vector<long long>   _numbers;
+    uint                     _max_size;
 
 };
